# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RAKT KAVACH - National Digital Blood Grid API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # USERS
...
  /analytics/state/{stateCode}:
    get:
      operationId: getStateAnalytics
      tags: [analytics]
      summary: Get state analytics
      parameters:
        - name: stateCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: State analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateAnalytics"
  /analytics/district/{districtCode}:
    get:
      operationId: getDistrictAnalytics
      tags: [analytics]
      summary: Get district analytics
      parameters:
        - name: districtCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: District analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DistrictAnalytics"
  /analytics/supply-demand:
    get:
      operationId: getSupplyDemandTrend
      tags: [analytics]
      summary: Get supply vs demand trend
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        "200":
          description: Supply demand trend
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyDemandTrend"
  /analytics/emergency-trends:
    get:
      operationId: getEmergencyTrends
      tags: [analytics]
      summary: Get emergency SOS trends
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
      responses:
        "200":
          description: Emergency trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyTrend"

  # DASHBOARD
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [CITIZEN, DONOR, PATIENT, HOSPITAL, BLOOD_BANK, LABORATORY, AMBULANCE, VOLUNTEER, DISTRICT_ADMIN, STATE_ADMIN, NATIONAL_ADMIN, SUPER_ADMIN]
        phone:
          type: string
        state:
          type: string
        district:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        isActive:
          type: boolean
        isVerified:
          type: boolean
        createdAt:
          type: string
...
    BloodGroupCount:
      type: object
      required: [bloodGroup, units, percentage]
      properties:
        bloodGroup:
          type: string
        units:
          type: number
        percentage:
          type: number
        status:
          type: string
          enum: [CRITICAL, LOW, ADEQUATE, SURPLUS]
...
    StateBloodData:
      type: object
      required: [stateCode, stateName, totalUnits, donors, hospitals, bloodBanks, activeEmergencies, supplyStatus]
      properties:
        stateCode:
          type: string
        stateName:
          type: string
        totalUnits:
          type: number
        donors:
          type: integer
        hospitals:
          type: integer
        bloodBanks:
          type: integer
        activeEmergencies:
          type: integer
        supplyStatus:
          type: string
          enum: [CRITICAL, LOW, ADEQUATE, SURPLUS]
...
    StateAnalytics:
      type: object
      required: [stateCode, stateName, totalUnits, donors, hospitals, bloodBanks, activeEmergencies, byBloodGroup, districtBreakdown]
      properties:
        stateCode:
          type: string
        stateName:
          type: string
        totalUnits:
          type: number
        donors:
          type: integer
        hospitals:
          type: integer
        bloodBanks:
          type: integer
        activeEmergencies:
          type: integer
        byBloodGroup:
          type: array
          items:
            $ref: "#/components/schemas/BloodGroupCount"
        districtBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/StateBloodData"
    DistrictAnalytics:
      type: object
      required: [districtCode, districtName, stateName, totalUnits, donors, hospitals, bloodBanks, activeEmergencies, byBloodGroup]
      properties:
        districtCode:
          type: string
        districtName:
          type: string
        stateName:
          type: string
        totalUnits:
          type: number
        donors:
          type: integer
        hospitals:
          type: integer
        bloodBanks:
          type: integer
        activeEmergencies:
          type: integer
        byBloodGroup:
          type: array
          items:
            $ref: "#/components/schemas/BloodGroupCount"
    TrendDataPoint:
      type: object
      required: [date, supply, demand]
      properties:
        date:
          type: string
        supply:
          type: number
        demand:
          type: number
    SupplyDemandTrend:
      type: object
      required: [period, data, supplyTotal, demandTotal, surplus]
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/TrendDataPoint"
        supplyTotal:
          type: number
        demandTotal:
          type: number
        surplus:
          type: number
    EmergencyTrendPoint:
      type: object
      required: [date, count, resolved, avgResolutionMinutes]
      properties:
        date:
          type: string
        count:
          type: integer
        resolved:
          type: integer
        avgResolutionMinutes:
          type: number
    EmergencyTrend:
      type: object
      required: [period, data, totalEmergencies, resolvedCount, avgResolutionTime]
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmergencyTrendPoint"
        totalEmergencies:
          type: integer
        resolvedCount:
          type: integer
        avgResolutionTime:
          type: number

    # DASHBOARD SCHEMAS
...