# lib/api-zod/src/generated/types/emergencyTrend.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RAKT KAVACH - National Digital Blood Grid API
 * OpenAPI spec version: 0.1.0
 */
import type { EmergencyTrendPoint } from './emergencyTrendPoint';

export interface EmergencyTrend {
  period: string;
  data: EmergencyTrendPoint[];
  totalEmergencies: number;
  resolvedCount: number;
  avgResolutionTime: number;
}