# lib/api-zod/src/generated/types/emergencyTrendPoint.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RAKT KAVACH - National Digital Blood Grid API
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyTrendPoint {
  date: string;
  count: number;
  resolved: number;
  avgResolutionMinutes: number;
}