# lib/api-zod/src/generated/types/supplyDemandTrend.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * RAKT KAVACH - National Digital Blood Grid API
 * OpenAPI spec version: 0.1.0
 */
import type { TrendDataPoint } from './trendDataPoint';

export interface SupplyDemandTrend {
  period: string;
  data: TrendDataPoint[];
  supplyTotal: number;
  demandTotal: number;
  surplus: number;
}